chore: replace gopsutil fork with upstream where possible
