+++ pkg/process/checks/system_info.go
@@ -8,9 +8,8 @@
 package checks
 
 import (
-	// waiting for upstream fix to platform collection in containerized environments
-	"github.com/DataDog/gopsutil/host"
 	"github.com/shirou/gopsutil/v4/cpu"
+	"github.com/shirou/gopsutil/v4/host"
 	"github.com/shirou/gopsutil/v4/mem"
 
 	model "github.com/DataDog/agent-payload/v5/process"
