+++ pkg/process/checks/system_info_darwin.go
@@ -11,10 +11,8 @@ import (
 	"errors"
 
 	model "github.com/DataDog/agent-payload/v5/process"
-	// difference between methods for collecting macOS platform, kernel version
-	// between shirou and Datadog psutil
-	"github.com/DataDog/gopsutil/host"
 	"github.com/shirou/gopsutil/v4/cpu"
+	"github.com/shirou/gopsutil/v4/host"
 	"github.com/shirou/gopsutil/v4/mem"
 	"golang.org/x/sys/unix"
 )
